loop game, menu, intro, game, lose, menu..

+++ Assets/MyStuff/Scripts/Data/Data.cs
@@ -1,6 +1,6 @@
 public static class Data
 {
     public static int currency             = 0 ;
-    public static int maxSeconds           = 0 ;
+    public static float maxSeconds           = 0 ;
     public static float currentTimer         = 0 ;
 }
